refactor(bonsai-archive): move ArchiveNodeKey into bonsai.archive package
