+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/trie/pathbased/bonsai/archive/ArchiveNodeKey.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.ethereum.trie.pathbased.bonsai.storage.triehistory;
+package org.hyperledger.besu.ethereum.trie.pathbased.bonsai.archive;
 
 import org.apache.tuweni.bytes.Bytes;
